First commit of Tumor vs Normal Pathway enrichment
We input saved Res objects from the DEseq2 software. The pathway enrichment code will look for significantly differentially expressed genes between tumor and normal for each cancer type. Following filtering the DE genes we use the clusterProfiler package to search for an enrichment within pathways or locations in the cell.